Util Files Initial Upload
The file SQLStatementUtils is used to populate Object containers with
results from a database. This file can return single elements, lists of
data transfer objects, or lists of Maps.

The Utils file provides some key functions that allow for the results to
be mapped correctly, cast to the expected input, and convert data into a
usable JSON format.